Abstract Class for Level Set Manifolds

Class for manifolds embedded in a vector space by a submersion.

Author

Nicolas Guigui and Nina Miolane

Super classes

rgeomstats::PythonClass -> rgeomstats::Manifold -> LevelSet

Public fields

embedding_space

An object of class VectorSpace specifying the embedding space.

embedding_metric

???

submersion

???

value

???

tangent_submersion

???

Methods

Public methods

LevelSet$new()

LevelSet$intrinsic_to_extrinsic_coords()

LevelSet$extrinsic_to_intrinsic_coords()

LevelSet$projection()

LevelSet$clone()

Inherited methods

rgeomstats::PythonClass$get_python_class()

rgeomstats::PythonClass$set_python_class()

rgeomstats::Manifold$belongs()

rgeomstats::Manifold$is_tangent()

rgeomstats::Manifold$random_point()

rgeomstats::Manifold$random_tangent_vec()

rgeomstats::Manifold$regularize()

rgeomstats::Manifold$set_metric()

rgeomstats::Manifold$to_tangent()

Method new()

The LevelSet class constructor.

Usage

LevelSet$new(
  dim,
  embedding_space,
  submersion,
  value,
  tangent_submersion,
  default_coords_type = "intrinsic",
  ...,
  py_cls = NULL
)

Arguments

dim

An integer value specifying the dimension of the manifold.

embedding_space

An object of class VectorSpace specifying the embedding space.

submersion

???

value

???

tangent_submersion

???

default_coords_type

A string specifying the coordinate type. Choices are extrinsic or intrinsic. Defaults to intrinsic.

...

Extra arguments to be passed to parent class constructors. See Manifold class.

py_cls

A Python object of class LevelSet. Defaults to NULL in which case it is instantiated on the fly using the other input arguments.

Returns

An object of class LevelSet.

Method intrinsic_to_extrinsic_coords()

Converts from intrinsic to extrinsic coordinates.

Usage

LevelSet$intrinsic_to_extrinsic_coords(point_intrinsic)

Arguments

point_intrinsic

A numeric array of shape dim specifying a point in the embedded manifold in intrinsic coordinates.

Returns

A numeric array of shape dim_embedding representing the same point in the embedded manifold in extrinsic coordinates.

Method extrinsic_to_intrinsic_coords()

Converts from extrinsic to intrinsic coordinates.

Usage

LevelSet$extrinsic_to_intrinsic_coords(point_extrinsic)

Arguments

point_extrinsic

A numeric array of shape dim_embedding specifying a point in the embedded manifold in extrinsic coordinates, i.E. in the coordinates of the embedding manifold.

Returns

A numeric array of shape dim representing the same point in the embedded manifold in intrinsic coordinates.

Method projection()

Projects a point in embedding manifold on embedded manifold.

Usage

LevelSet$projection(point)

Arguments

point

A numeric array of shape dim_embedding specifying a point in the embedding manifold.

Returns

A numeric array of shape dim_embedding storing the projected point.

Method clone()

The objects of this class are cloneable with this method.

Usage

LevelSet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.