Abstract Class for Manifolds

An R6::R6Class object implementing the base Manifold class. In other words, a topological space that locally resembles Euclidean space near each point.

Author

Nina Miolane

Super class

rgeomstats::PythonClass -> Manifold

Public fields

dim

An integer value specifying the dimension of the manifold.

shape

An integer vector specifying the shape of one element of the manifold. Defaults to NULL.

metric

A RiemannianMetric object specifying the metric to use on the manifold. Defaults to NULL.

default_coords_type

A string specifying the coordinate type. Choices are extrensic or intrinsic. Dedaults to intrinsic.

default_point_type

A string specifying the point type. Choices are vector or matrix. It is automatically determined depending on the manifold.

Methods

Public methods

Manifold$new()

Manifold$belongs()

Manifold$is_tangent()

Manifold$to_tangent()

Manifold$random_point()

Manifold$regularize()

Manifold$set_metric()

Manifold$random_tangent_vec()

Manifold$clone()

Inherited methods

rgeomstats::PythonClass$get_python_class()

rgeomstats::PythonClass$set_python_class()

Method new()

The Manifold class constructor.

Usage

Manifold$new(
  dim,
  shape = NULL,
  metric = NULL,
  default_coords_type = "intrinsic"
)

Arguments

dim

An integer value specifying the dimension of the manifold.

shape

An integer vector specifying the shape of one element of the manifold. Defaults to NULL.

metric

A RiemannianMetric object specifying the metric to use on the manifold. Defaults to NULL.

default_coords_type

A string specifying the coordinate type. Choices are extrensic or intrinsic. Defaults to intrinsic.

Returns

An object of class Manifold.

Method belongs()

Evaluates if a point belongs to the manifold.

Usage

Manifold$belongs(point, atol = gs$backend$atol)

Arguments

point

An numeric array of shape dim specifying a point to be checked.

atol

A numeric value specifying the absolute tolerance for checking. Defaults to gs$backend$atol.

Returns

A boolean that tells whether the input point belongs to the manifold.

Examples

if (reticulate::py_module_available("geomstats")) {
  mf <- SPDMatrices$new(n = 3)
  A <- diag(1, 3)
  mf$belongs(A)
  B <- diag(-1, 3)
  mf$belongs(B)
}

Method is_tangent()

Checks whether a vector is tangent at a base point.

Usage

Manifold$is_tangent(vector, base_point = NULL, atol = gs$backend$atol)

Arguments

vector

An numeric array of shape dim specifying a vector to be checked.

base_point

An numeric array of shape dim specifying a base point which belongs to the manifold.

atol

A numeric value specifying the absolute tolerance for checking. Defaults to gs$backend$atol.

Returns

A boolean that tells whether the input vector is ≥tangent to the manifold at base_point.

Method to_tangent()

Projects a vector to a tangent space of the manifold.

Usage

Manifold$to_tangent(vector, base_point = NULL)

Arguments

vector

A (list of) numeric array(s) of shape dim specifying a (list of) vector(s) to project on the manifold.

base_point

A (list of) numeric array(s) of shape dim specifying a (list of) point(s) on the manifold. Defaults to NULL.

Returns

A (list of) numeric array(s) of shape dim storing a (list of) projected vector(s) that belong to the manifold.

Method random_point()

Samples random points on the manifold.

Usage

Manifold$random_point(n_samples = 1, bound = 1)

Arguments

n_samples

An integer value specifying the number of samples to be drawn. Defaults to 1L.

bound

A numeric value specifying the bound of the interval in which to sample for non compact manifolds. Defaults to 1L.

base_point

An numeric array of shape dim specifying a base point which belongs to the manifold.

Details

If the manifold is compact, a uniform distribution is used.

Returns

A list of numeric arrays representing a sample of points on the manifold.

Examples

if (reticulate::py_module_available("geomstats")) {
  mf <- SPDMatrices$new(n = 3)
  # mf$random_point(10) # TO DO: uncomment when bug fixed in gs
}

Method regularize()

Regularizes a point to the canonical representation for the manifold.

Usage

Manifold$regularize(point)

Arguments

point

A numeric array of shape dim specifying a point which belongs to the manifold.

Returns

A numeric array of shape dim storing a regularized version of the input point.

Method set_metric()

Sets the Riemannian Metric associated to the manifold.

Usage

Manifold$set_metric(metric)

Arguments

metric

An object of class RiemannianMetric.

Returns

The Manifold class itself invisibly.

Method random_tangent_vec()

Generates a random tangent vector.

Usage

Manifold$random_tangent_vec(base_point, n_samples = 1)

Arguments

base_point

An numeric array of shape dim specifying a base point which belongs to the manifold.

n_samples

An integer value specifying the number of samples to be drawn. Defaults to 1L.

Returns

A list of numeric arrays representing a sample of vectors that are tangent to the manifold at base_point.

Examples

if (reticulate::py_module_available("geomstats")) {
  mf <- SPDMatrices$new(n = 3)
  mf$random_tangent_vec(diag(1, 3), 10)
}

Method clone()

The objects of this class are cloneable with this method.

Usage

Manifold$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `Manifold$belongs`
## ------------------------------------------------

if (reticulate::py_module_available("geomstats")) {
  mf <- SPDMatrices$new(n = 3)
  A <- diag(1, 3)
  mf$belongs(A)
  B <- diag(-1, 3)
  mf$belongs(B)
}

## ------------------------------------------------
## Method `Manifold$random_point`
## ------------------------------------------------

if (reticulate::py_module_available("geomstats")) {
  mf <- SPDMatrices$new(n = 3)
  # mf$random_point(10) # TO DO: uncomment when bug fixed in gs
}

## ------------------------------------------------
## Method `Manifold$random_tangent_vec`
## ------------------------------------------------

if (reticulate::py_module_available("geomstats")) {
  mf <- SPDMatrices$new(n = 3)
  mf$random_tangent_vec(diag(1, 3), 10)
}