Class for the Manifold of Symmetric Positive Definite Matrices

This function generates an instance of the class for the manifold of symmetric positive definite matrices \(\mathrm{SPD}(n)\).

Usage

SPDMatrix(n, ...)

Arguments

n

An integer value specifying the number of rows and columns of the matrices.

...

Extra arguments to be passed to parent class constructors. See OpenSet and Manifold classes.

Value

An object of class SPDMatrices.

See also

Author

Yann Thanwerdas

Examples

if (reticulate::py_module_available("geomstats")) {
  spd3 <- SPDMatrix(n = 3)
  spd3
}