Abstract Class for Vector Space Manifolds

Abstract class for vector spaces.

Author

Nicolas Guigui and Nina Miolane

Super classes

rgeomstats::PythonClass -> rgeomstats::Manifold -> VectorSpace

Public fields

basis

Basis of the vector space.

Methods

Public methods

VectorSpace$new()

VectorSpace$projection()

VectorSpace$clone()

Inherited methods

rgeomstats::PythonClass$get_python_class()

rgeomstats::PythonClass$set_python_class()

rgeomstats::Manifold$belongs()

rgeomstats::Manifold$is_tangent()

rgeomstats::Manifold$random_point()

rgeomstats::Manifold$random_tangent_vec()

rgeomstats::Manifold$regularize()

rgeomstats::Manifold$set_metric()

rgeomstats::Manifold$to_tangent()

Method new()

The VectorSpace class constructor.

Usage

VectorSpace$new(shape, ..., py_cls = NULL)

Arguments

shape

An integer vector specifying the shape of one element of the manifold. Defaults to NULL.

...

Extra arguments to be passed to parent class constructors. See Manifold class.

py_cls

A Python object of class VectorSpace. Defaults to NULL in which case it is instantiated on the fly using the other input arguments.

Returns

An object of class VectorSpace.

Method projection()

Project a point onto the vector space.

Usage

VectorSpace$projection(point)

Arguments

point

A numeric array of shape dim specifying a vector in the ambient space onto the manifold.

Details

This method is for compatibility and returns point. point should have the right shape.

Returns

A numeric array of shape dim storing the input vector projected onto the manifold.

Method clone()

The objects of this class are cloneable with this method.

Usage

VectorSpace$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.