Manifold of Symmetric Positive Definite Matrices

Class for the manifold of symmetric positive definite (SPD) matrices.

Author

Yann Thanwerdas

Methods

Public methods

SPDMatrices$new()

SPDMatrices$belongs()

SPDMatrices$cholesky_factor()

SPDMatrices$differential_cholesky_factor()

SPDMatrices$differential_exp()

SPDMatrices$differential_log()

SPDMatrices$differential_power()

SPDMatrices$clone()

Method new()

The SPDMatrices class constructor.

Usage

SPDMatrices$new(n)

Arguments

n

An integer value representing the shape of the n x n matrices.

Returns

A SPDMatrices R6::R6Class object.

Examples

if (reticulate::py_module_available("geomstats")) {
  spdm <- SPDMatrices$new(n = 3)
  spdm
}

Method belongs()

Check if a matrix is symmetric with positive eigenvalues.

Usage

SPDMatrices$belongs(mat, atol = 1e-12)

Arguments

mat

A numeric matrix to be checked.

atol

A numeric value specifying the absolute tolerance for checking. Defaults to 1e-12.

Returns

A boolean that tells whether the input matrix is SPD.

Examples

if (reticulate::py_module_available("geomstats")) {
  spdm <- SPDMatrices$new(n = 3)
  A <- diag(1, 3)
  spdm$belongs(A)
  B <- diag(-1, 3)
  spdm$belongs(B)
}

Method cholesky_factor()

Computes Cholesky factor for a symmetric positive definite matrix.

Usage

SPDMatrices$cholesky_factor(mat)

Arguments

mat

An SPD matrix.

Returns

The Cholesky factor represented as a numeric vector which stores its lower triangular matrix including the diagonal elements in a column-major fashion.

Examples

if (reticulate::py_module_available("geomstats")) {
  spdm <- SPDMatrices$new(n = 3)
  V <- cbind(
    c(sqrt(2) / 2, -sqrt(2) / 2, 0),
    c(sqrt(2) / 2, sqrt(2) / 2, 0),
    c(0, 0, 1)
  )
  A <- V %*% diag(1:3) %*% t(V)
  spdm$cholesky_factor(A)
}

Method differential_cholesky_factor()

Computes the differential of the Cholesky factor map.

Usage

SPDMatrices$differential_cholesky_factor(tangent_vec, base_point)

Arguments

tangent_vec

A symmetric matrix specifying the tangent vector at base point.

base_point

An SPD matrix specifying the base point.

Returns

The differential of the Cholesky factor map represented as a numeric vector which stores its lower triangular matrix including the diagonal elements in a column-major fashion.

Examples

if (reticulate::py_module_available("geomstats")) {
  spdm <- SPDMatrices$new(n = 3)
  V <- cbind(
    c(sqrt(2) / 2, -sqrt(2) / 2, 0),
    c(sqrt(2) / 2, sqrt(2) / 2, 0),
    c(0, 0, 1)
  )
  A <- V %*% diag(1:3) %*% t(V)
  spdm$differential_cholesky_factor(diag(1, 3), A)
}

Method differential_exp()

Computes the differential of the matrix exponential.

Usage

SPDMatrices$differential_exp(tangent_vec, base_point)

Arguments

tangent_vec

A symmetric matrix specifying the tangent vector at base point.

base_point

An SPD matrix specifying the base point.

Returns

A matrix storing the differential of the matrix exponential on SPD matrices at base_point applied to tangent_vec.

Examples

if (reticulate::py_module_available("geomstats")) {
  spdm <- SPDMatrices$new(n = 3)
  V <- cbind(
    c(sqrt(2) / 2, -sqrt(2) / 2, 0),
    c(sqrt(2) / 2, sqrt(2) / 2, 0),
    c(0, 0, 1)
  )
  A <- V %*% diag(1:3) %*% t(V)
  spdm$differential_exp(diag(1, 3), A)
}

Method differential_log()

Computes the differential of the matrix logarithm.

Usage

SPDMatrices$differential_log(tangent_vec, base_point)

Arguments

tangent_vec

A symmetric matrix specifying the tangent vector at base point.

base_point

An SPD matrix specifying the base point.

Returns

A matrix storing the differential of the matrix logarithm on SPD matrices at base_point applied to tangent_vec.

Examples

if (reticulate::py_module_available("geomstats")) {
  spdm <- SPDMatrices$new(n = 3)
  V <- cbind(
    c(sqrt(2) / 2, -sqrt(2) / 2, 0),
    c(sqrt(2) / 2, sqrt(2) / 2, 0),
    c(0, 0, 1)
  )
  A <- V %*% diag(1:3) %*% t(V)
  spdm$differential_log(diag(1, 3), A)
}

Method differential_power()

Computes the differential of the matrix power function.

Usage

SPDMatrices$differential_power(power, tangent_vec, base_point)

Arguments

power

An integer scalar specifying the desired power.

tangent_vec

A symmetric matrix specifying the tangent vector at base point.

base_point

An SPD matrix specifying the base point.

Returns

A matrix storing the differential of the power function on \(\mathrm{SPD}(n)\): $$A^p = \exp(p \log(A))$$ at base_point applied to tangent_vec.

Examples

if (reticulate::py_module_available("geomstats")) {
  spdm <- SPDMatrices$new(n = 3)
  V <- cbind(
    c(sqrt(2) / 2, -sqrt(2) / 2, 0),
    c(sqrt(2) / 2, sqrt(2) / 2, 0),
    c(0, 0, 1)
  )
  A <- V %*% diag(1:3) %*% t(V)
  spdm$differential_power(2, diag(1, 3), A)
}

Method clone()

The objects of this class are cloneable with this method.

Usage

SPDMatrices$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `SPDMatrices$new`
## ------------------------------------------------

if (reticulate::py_module_available("geomstats")) {
  spdm <- SPDMatrices$new(n = 3)
  spdm
}

## ------------------------------------------------
## Method `SPDMatrices$belongs`
## ------------------------------------------------

if (reticulate::py_module_available("geomstats")) {
  spdm <- SPDMatrices$new(n = 3)
  A <- diag(1, 3)
  spdm$belongs(A)
  B <- diag(-1, 3)
  spdm$belongs(B)
}

## ------------------------------------------------
## Method `SPDMatrices$cholesky_factor`
## ------------------------------------------------

if (reticulate::py_module_available("geomstats")) {
  spdm <- SPDMatrices$new(n = 3)
  V <- cbind(
    c(sqrt(2) / 2, -sqrt(2) / 2, 0),
    c(sqrt(2) / 2, sqrt(2) / 2, 0),
    c(0, 0, 1)
  )
  A <- V %*% diag(1:3) %*% t(V)
  spdm$cholesky_factor(A)
}

## ------------------------------------------------
## Method `SPDMatrices$differential_cholesky_factor`
## ------------------------------------------------

if (reticulate::py_module_available("geomstats")) {
  spdm <- SPDMatrices$new(n = 3)
  V <- cbind(
    c(sqrt(2) / 2, -sqrt(2) / 2, 0),
    c(sqrt(2) / 2, sqrt(2) / 2, 0),
    c(0, 0, 1)
  )
  A <- V %*% diag(1:3) %*% t(V)
  spdm$differential_cholesky_factor(diag(1, 3), A)
}

## ------------------------------------------------
## Method `SPDMatrices$differential_exp`
## ------------------------------------------------

if (reticulate::py_module_available("geomstats")) {
  spdm <- SPDMatrices$new(n = 3)
  V <- cbind(
    c(sqrt(2) / 2, -sqrt(2) / 2, 0),
    c(sqrt(2) / 2, sqrt(2) / 2, 0),
    c(0, 0, 1)
  )
  A <- V %*% diag(1:3) %*% t(V)
  spdm$differential_exp(diag(1, 3), A)
}

## ------------------------------------------------
## Method `SPDMatrices$differential_log`
## ------------------------------------------------

if (reticulate::py_module_available("geomstats")) {
  spdm <- SPDMatrices$new(n = 3)
  V <- cbind(
    c(sqrt(2) / 2, -sqrt(2) / 2, 0),
    c(sqrt(2) / 2, sqrt(2) / 2, 0),
    c(0, 0, 1)
  )
  A <- V %*% diag(1:3) %*% t(V)
  spdm$differential_log(diag(1, 3), A)
}

## ------------------------------------------------
## Method `SPDMatrices$differential_power`
## ------------------------------------------------

if (reticulate::py_module_available("geomstats")) {
  spdm <- SPDMatrices$new(n = 3)
  V <- cbind(
    c(sqrt(2) / 2, -sqrt(2) / 2, 0),
    c(sqrt(2) / 2, sqrt(2) / 2, 0),
    c(0, 0, 1)
  )
  A <- V %*% diag(1:3) %*% t(V)
  spdm$differential_power(2, diag(1, 3), A)
}