Class for the log-Euclidean Metric on the Manifold of Symmetric Positive Definite Matrices

An R6::R6Class object implementing the SPDMetricLogEuclidean class. This is the class for the log-Euclidean metric on the SPD manifold.

Author

Yann Thanwerdas

Super classes

rgeomstats::PythonClass -> rgeomstats::Connection -> rgeomstats::RiemannianMetric -> SPDMetricLogEuclidean

Public fields

n

An integer value specifying the shape of the matrices: \(n \times n\).

Methods

Public methods

SPDMetricLogEuclidean$new()

SPDMetricLogEuclidean$clone()

Inherited methods

rgeomstats::PythonClass$get_python_class()

rgeomstats::PythonClass$set_python_class()

rgeomstats::Connection$christoffels()

rgeomstats::Connection$curvature()

rgeomstats::Connection$curvature_derivative()

rgeomstats::Connection$directional_curvature()

rgeomstats::Connection$directional_curvature_derivative()

rgeomstats::Connection$exp()

rgeomstats::Connection$geodesic()

rgeomstats::Connection$geodesic_equation()

rgeomstats::Connection$injectivity_radius()

rgeomstats::Connection$ladder_parallel_transport()

rgeomstats::Connection$log()

rgeomstats::Connection$parallel_transport()

rgeomstats::RiemannianMetric$closest_neighbor_index()

rgeomstats::RiemannianMetric$cometric_matrix()

rgeomstats::RiemannianMetric$diameter()

rgeomstats::RiemannianMetric$dist()

rgeomstats::RiemannianMetric$dist_broadcast()

rgeomstats::RiemannianMetric$dist_pairwise()

rgeomstats::RiemannianMetric$hamiltonian()

rgeomstats::RiemannianMetric$inner_coproduct()

rgeomstats::RiemannianMetric$inner_product()

rgeomstats::RiemannianMetric$inner_product_derivative_matrix()

rgeomstats::RiemannianMetric$metric_matrix()

rgeomstats::RiemannianMetric$norm()

rgeomstats::RiemannianMetric$normal_basis()

rgeomstats::RiemannianMetric$normalize()

rgeomstats::RiemannianMetric$random_unit_tangent_vec()

rgeomstats::RiemannianMetric$sectional_curvature()

rgeomstats::RiemannianMetric$squared_dist()

rgeomstats::RiemannianMetric$squared_norm()

Method new()

The SPDMetricLogEuclidean class constructor.

Usage

SPDMetricLogEuclidean$new(n)

Arguments

n

An integer value specifying the shape of the matrices: \(n \times n\).

Returns

An object of class SPDMetricLogEuclidean.

Examples

if (reticulate::py_module_available("geomstats")) {
  mt <- SPDMetricLogEuclidean$new(n = 3)
  mt
}

Method clone()

The objects of this class are cloneable with this method.

Usage

SPDMetricLogEuclidean$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `SPDMetricLogEuclidean$new`
## ------------------------------------------------

if (reticulate::py_module_available("geomstats")) {
  mt <- SPDMetricLogEuclidean$new(n = 3)
  mt
}